fix webcontent events, move views to its own folder
