+++ src/main/views/modalManager.js
@@ -5,8 +5,9 @@ import {ipcMain} from 'electron';
 
 import {RETRIEVE_MODAL_INFO, MODAL_CANCEL, MODAL_RESULT, MODAL_OPEN, MODAL_CLOSE} from 'common/communication.js';
 
+import * as WindowManager from '../windows/windowManager';
+
 import {ModalView} from './modalView';
-import * as WindowManager from './windows/windowManager';
 
 let modalQueue = [];
