+++ src/main/views/viewManager.js
@@ -1,40 +1,24 @@
 // Copyright (c) 2016-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 import log from 'electron-log';
-import {BrowserView, BrowserWindow, dialog, shell} from 'electron';
+import {BrowserView, dialog} from 'electron';
 
-import {SECOND, DEVELOPMENT, PRODUCTION} from 'common/utils/constants';
+import {SECOND} from 'common/utils/constants';
 import {UPDATE_TARGET_URL, FOUND_IN_PAGE, SET_SERVER_KEY, LOAD_SUCCESS, LOAD_FAILED} from 'common/communication';
 import urlUtils from 'common/utils/url';
-import Utils from 'common/utils/util';
 
-import {protocols} from '../../electron-builder.json';
+import contextMenu from '../contextMenu';
+import {MattermostServer} from '../MattermostServer';
+import {getLocalURLString, getLocalPreload} from '../utils';
 
-import contextMenu from './contextMenu';
-import {MattermostServer} from './MattermostServer';
 import {MattermostView} from './MattermostView';
-import {getLocalURLString, getLocalPreload} from './utils';
 import {showModal} from './modalManager';
-import allowProtocolDialog from './allowProtocolDialog';
+import {addWebContentsEventListeners} from './webContentEvents';
 
 const URL_VIEW_DURATION = 10 * SECOND;
 const URL_VIEW_HEIGHT = 36;
 const FINDER_WIDTH = 310;
 const FINDER_HEIGHT = 40;
-
-const nixUA = 'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome Safari/537.36';
-
-const popupUserAgent = {
-    darwin: 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_14_6) AppleWebKit/537.36 (KHTML, like Gecko) Chrome Safari/537.36',
-    win32: 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome Safari/537.36',
-    aix: nixUA,
-    freebsd: nixUA,
-    linux: nixUA,
-    openbsd: nixUA,
-    sunos: nixUA,
-};
-
-const scheme = protocols && protocols[0] && protocols[0].schemes && protocols[0].schemes[0];
 
 export class ViewManager {
     constructor(config) {
@@ -43,209 +27,11 @@ export class ViewManager {
         this.views = new Map(); // keep in mind that this doesn't need to hold server order, only tabs on the renderer need that.
         this.currentView = null;
         this.urlView = null;
-
-        // tracking in progress custom logins
-        this.customLogins = {};
-
-        // hold listener removal functions
-        this.listeners = {};
-
-        this.popupWindow = null;
     }
 
-    // temporarily added here
-
-    isTrustedPopupWindow = (webContents) => {
-        if (!webContents) {
-            return false;
-        }
-        if (!this.popupWindow) {
-            return false;
-        }
-        return BrowserWindow.fromWebContents(webContents) === this.popupWindow;
+    getServers = () => {
+        return this.configServers;
     }
-
-    willNavigate = (_event, url) => {
-        const contentID = event.sender.id;
-        const parsedURL = urlUtils.parseURL(url);
-        const server = urlUtils.getServer(parsedURL, this.configServers);
-
-        if (server && (urlUtils.isTeamUrl(server.url, parsedURL) || urlUtils.isAdminUrl(server.url, parsedURL) || this.isTrustedPopupWindow(event.sender))) {
-            return;
-        }
-
-        if (urlUtils.isCustomLoginURL(parsedURL, server, this.configServers)) {
-            return;
-        }
-        if (parsedURL.protocol === 'mailto:') {
-            return;
-        }
-        if (this.customLogins[contentID].inProgress) {
-            return;
-        }
-        const mode = Utils.runMode();
-        if (((mode === DEVELOPMENT || mode === PRODUCTION) &&
-        (parsedURL.path === 'renderer/index.html' || parsedURL.path === 'renderer/settings.html'))) {
-            log.info('loading settings page');
-            return;
-        }
-
-        log.info(`Prevented desktop from navigating to: ${url}`);
-        event.preventDefault();
-    };
-
-    didStartNavigation = (event, url) => {
-        const contentID = event.sender.id;
-        const parsedURL = urlUtils.parseURL(url);
-        const server = urlUtils.getServer(parsedURL, this.configServers);
-
-        if (!urlUtils.isTrustedURL(parsedURL, this.configServers)) {
-            return;
-        }
-
-        if (urlUtils.isCustomLoginURL(parsedURL, server, this.configServers)) {
-            this.customLogins[contentID].inProgress = true;
-        } else if (this.customLogins[contentID].inProgress) {
-            this.customLogins[contentID].inProgress = false;
-        }
-    };
-
-    newWindow = (event, url) => {
-        const parsedURL = urlUtils.parseURL(url);
-
-        // Dev tools case
-        if (parsedURL.protocol === 'devtools:') {
-            return;
-        }
-        event.preventDefault();
-
-        // Check for valid URL
-        if (!urlUtils.isValidURI(url)) {
-            return;
-        }
-
-        // Check for custom protocol
-        if (parsedURL.protocol !== 'http:' && parsedURL.protocol !== 'https:' && parsedURL.protocol !== `${scheme}:`) {
-            allowProtocolDialog.handleDialogEvent(parsedURL.protocol, url);
-            return;
-        }
-
-        const server = urlUtils.getServer(parsedURL, this.configServers);
-
-        if (!server) {
-            shell.openExternal(url);
-            return;
-        }
-
-        // Public download links case
-        // TODO: We might be handling different types differently in the future, for now
-        // we are going to mimic the browser and just pop a new browser window for public links
-        if (parsedURL.pathname.match(/^(\/api\/v[3-4]\/public)*\/files\//)) {
-            shell.openExternal(url);
-            return;
-        }
-
-        if (parsedURL.pathname.match(/^\/help\//)) {
-            // Help links case
-            // continue to open special case internal urls in default browser
-            shell.openExternal(url);
-            return;
-        }
-
-        if (urlUtils.isTeamUrl(server.url, parsedURL, true)) {
-            log.info(`${url} is a known team, preventing to open a new window`);
-            return;
-        }
-        if (urlUtils.isAdminUrl(server.url, parsedURL)) {
-            log.info(`${url} is an admin console page, preventing to open a new window`);
-            return;
-        }
-        if (this.popupWindow && !this.popupWindow.closed && this.popupWindow.getURL() === url) {
-            log.info(`Popup window already open at provided url: ${url}`);
-            return;
-        }
-
-        // TODO: move popups to its own and have more than one.
-        if (urlUtils.isPluginUrl(server.url, parsedURL) || urlUtils.isManagedResource(server.url, parsedURL)) {
-            if (!this.popupWindow || this.popupWindow.closed) {
-                this.popupWindow = new BrowserWindow({
-                    backgroundColor: '#fff', // prevents blurry text: https://electronjs.org/docs/faq#the-font-looks-blurry-what-is-this-and-what-can-i-do
-                    //parent: WindowManager.getMainWindow(),
-                    show: false,
-                    center: true,
-                    webPreferences: {
-                        nodeIntegration: false,
-                        contextIsolation: true,
-                        spellcheck: (typeof this.viewOptions.spellcheck === 'undefined' ? true : this.viewOptions.spellcheck),
-                    },
-                });
-                this.popupWindow.once('ready-to-show', () => {
-                    this.popupWindow.show();
-                });
-                this.popupWindow.once('closed', () => {
-                    this.popupWindow = null;
-                });
-            }
-
-            if (urlUtils.isManagedResource(server.url, parsedURL)) {
-                this.popupWindow.loadURL(url);
-            } else {
-                // currently changing the userAgent for popup windows to allow plugins to go through google's oAuth
-                // should be removed once a proper oAuth2 implementation is setup.
-                this.popupWindow.loadURL(url, {
-                    userAgent: popupUserAgent[process.platform],
-                });
-            }
-        }
-    };
-
-    removeWebContentsListeners = (id) => {
-        if (this.listeners[id]) {
-            this.listeners[id]();
-        }
-    }
-
-    handleAppWebContentsCreated = (contents) => {
-        // initialize custom login tracking
-        this.customLogins[contents.id] = {
-            inProgress: false,
-        };
-
-        if (this.listeners[contents.id]) {
-            this.removeWebContentsListeners(contents.id);
-        }
-
-        contents.on('will-navigate', this.willNavigate);
-
-        // handle custom login requests (oath, saml):
-        // 1. are we navigating to a supported local custom login path from the `/login` page?
-        //    - indicate custom login is in progress
-        // 2. are we finished with the custom login process?
-        //    - indicate custom login is NOT in progress
-        contents.on('did-start-navigation', this.didStartNavigation);
-
-        contents.on('new-window', this.newWindow);
-
-        const removeListeners = () => {
-            try {
-                contents.removeListener('will-navigate', this.willNavigate);
-                contents.removeListener('did-start-navigation', this.didStartNavigation);
-                contents.removeListener('new-window', this.newWindow);
-            } catch (e) {
-                log.error(`Error while trying to detach listeners, this might be ok if the process crashed: ${e}`);
-            }
-        };
-
-        this.listeners[contents.id] = removeListeners;
-        contents.once('render-process-gone', (event, details) => {
-            if (details !== 'clean-exit') {
-                log.error(`Renderer process for a webcontent is no longer available: ${details}`);
-            }
-            removeListeners();
-        });
-    }
-
-    // end of temporarily added here
 
     loadServer = (server, mainWindow) => {
         const srv = new MattermostServer(server.name, server.url);
@@ -339,7 +125,7 @@ export class ViewManager {
             this.showByName(this.currentView);
         }
         const view = this.views.get(viewName);
-        this.handleAppWebContentsCreated(view.view.webContents);
+        addWebContentsEventListeners(view, this.getServers);
     }
 
     getCurrentView() {
